requirements.txt
django
selenium
fabric
djangorestframework
djangorestframework-jwt
django-cors-headers
psycopg2
django-contrib-comments
djangorestframework
